TOPR/requirements.txt
django==1.9
kafka-python
